Refs #19402: Address compiler warnings

+++ src/cpp/rtps/participant/RTPSParticipantImpl.cpp
@@ -2795,7 +2795,7 @@ bool RTPSParticipantImpl::fill_discovery_data_from_cdr_message(
     serialized_msg.wraps = true;
 
     serialized_msg.buffer = msg.value().entity_proxy().data();
-    serialized_msg.length = msg.value().entity_proxy().size();
+    serialized_msg.length = static_cast<uint32_t>(msg.value().entity_proxy().size());
 
     ret = data.readFromCDRMessage(
         &serialized_msg,
@@ -2816,7 +2816,7 @@ bool RTPSParticipantImpl::fill_discovery_data_from_cdr_message(
     serialized_msg.wraps = true;
 
     serialized_msg.buffer = msg.value().entity_proxy().data();
-    serialized_msg.length = msg.value().entity_proxy().size();
+    serialized_msg.length = static_cast<uint32_t>(msg.value().entity_proxy().size());
 
     ret = data.readFromCDRMessage(
         &serialized_msg,
@@ -2836,7 +2836,7 @@ bool RTPSParticipantImpl::fill_discovery_data_from_cdr_message(
     serialized_msg.wraps = true;
 
     serialized_msg.buffer = msg.value().entity_proxy().data();
-    serialized_msg.length = msg.value().entity_proxy().size();
+    serialized_msg.length = static_cast<uint32_t>(msg.value().entity_proxy().size());
 
     ret = data.readFromCDRMessage(
         &serialized_msg,
